shiv-stock-dashboard/requirements.txt
Django>=4.2,<5.0
python-dotenv>=1.0.0
yfinance>=0.2.36
numpy>=1.24.0
pandas>=2.0.0
